remove Spawnable SOs
