--- UnityProject/Assets/Resources/ScriptableObjects/Lifecycle/Spawnable.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 750f797ecd385a547bba1aa20aeea4c5
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
