--- UnityProject/Assets/Resources/ScriptableObjects/Lifecycle/Spawnable/Prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bc716a82ecd699d4d88d5f5e80ae072a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
