--- UnityProject/Assets/Resources/ScriptableObjects/Lifecycle/Spawnable/Prefab/hydroponics.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4d3d9f9ee94e9ad4d971232d37b716d1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
