--- UnityProject/Assets/Resources/ScriptableObjects/Lifecycle/Spawnable/Prefab/hydroponics/cultivator.asset
@@ -1,15 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e50d9d7a542f468b948bbaa80c532f11, type: 3}
-  m_Name: cultivator
-  m_EditorClassIdentifier: 
-  Prefab: {fileID: 1519710982558066, guid: a6ad762962da51448b30dea32427e068, type: 3}
