--- UnityProject/Assets/Resources/ScriptableObjects/Lifecycle/Spawnable/Prefab/hydroponics/cultivator.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ab89f009663b6b24684d56a90d4fbb48
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
