--- UnityProject/Assets/Resources/ScriptableObjects/Lifecycle/Spawnable/Prefab/hydroponics/trowel.asset
@@ -1,15 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e50d9d7a542f468b948bbaa80c532f11, type: 3}
-  m_Name: trowel
-  m_EditorClassIdentifier: 
-  Prefab: {fileID: 1519710982558066, guid: 549f7d7203498874bafc385496a54fb5, type: 3}
