--- UnityProject/Assets/Resources/ScriptableObjects/Lifecycle/Spawnable/Prefab/hydroponics/trowel.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2350d1d6afe5cce4081a9d362a3312df
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
