+++ UnityProject/Assets/Resources/ScriptableObjects/Lifecycle/SpawnableLists/Closets/hydroponics.asset
@@ -15,10 +15,10 @@ MonoBehaviour:
   contents:
   - {fileID: 3011710637427380133, guid: 489cb478de43746cda8af4fd63be9282, type: 3}
   - {fileID: 3011710637427380133, guid: 489cb478de43746cda8af4fd63be9282, type: 3}
-  spawnables:
-  - {fileID: 11400000, guid: ab89f009663b6b24684d56a90d4fbb48, type: 2}
-  - {fileID: 11400000, guid: ab89f009663b6b24684d56a90d4fbb48, type: 2}
-  - {fileID: 11400000, guid: 2350d1d6afe5cce4081a9d362a3312df, type: 2}
-  - {fileID: 11400000, guid: 2350d1d6afe5cce4081a9d362a3312df, type: 2}
-  - {fileID: 11400000, guid: 9064b128c1367e64d9d142db1ff87fb5, type: 2}
-  - {fileID: 11400000, guid: 9064b128c1367e64d9d142db1ff87fb5, type: 2}
+  - {fileID: 1519710982558066, guid: a6ad762962da51448b30dea32427e068, type: 3}
+  - {fileID: 1519710982558066, guid: a6ad762962da51448b30dea32427e068, type: 3}
+  - {fileID: 1519710982558066, guid: 549f7d7203498874bafc385496a54fb5, type: 3}
+  - {fileID: 1519710982558066, guid: 549f7d7203498874bafc385496a54fb5, type: 3}
+  - {fileID: 1801347380884054, guid: 2b30815126120534d823a2f2fdf0a26d, type: 3}
+  - {fileID: 1801347380884054, guid: 2b30815126120534d823a2f2fdf0a26d, type: 3}
+  spawnables: []
